requirements.txt
kivymd>=0.104.1
python-bidi
persiantools